sorted subpages alphabetically
