+++ package.json
@@ -14,6 +14,7 @@
     "gravatar": "^1.8.0",
     "jsonwebtoken": "^8.4.0",
     "jwt-decode": "^2.2.0",
+    "lodash": "^4.17.11",
     "mongoose": "^5.3.13",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
